+++ tests/Tests/Container/fixtures/ARecursive.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Container\Fixtures;
+
+class ARecursive
+{
+	public BRecursive $b;
+
+	public function __construct(BRecursive $b)
+	{
+		$this->b = $b;
+	}
+}