+++ tests/Tests/Container/fixtures/BRecursive.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Container\Fixtures;
+
+class BRecursive
+{
+	public ARecursive $a;
+
+	public function __construct(ARecursive $a)
+	{
+		$this->a = $a;
+	}
+}