+++ tests/Tests/Container/fixtures/C.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Container\Fixtures;
+
+class C
+{
+	public B $b;
+
+	public function __construct(B $b)
+	{
+		$this->b = $b;
+	}
+}