+++ tests/Tests/Container/fixtures/E.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Container\Fixtures;
+
+class E
+{
+	public string $message;
+
+	public ?A $a;
+
+	public function __construct(string $message = '', ?A $a = null)
+	{
+		$this->message = $message;
+		$this->a       = $a;
+	}
+}