+++ tests/Tests/Container/fixtures/F.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Container\Fixtures;
+
+class F
+{
+	public string $message;
+
+	public function __construct(string $message)
+	{
+		$this->message = $message;
+	}
+}